Successfully replaced main functions with CALCKR
Successfully replaced the function in MAIN() responsible for constructing residual and tangent with a single function CALCKR(), which contains all of those functions. Had to figure out which local variables to pass into CALCKR(). Output is exactly the same as before change, but now residual and tangent functions are entirely contained in CALCKR(), which will be used for line search later.