Regula falsi line search. LV sim promising
Implemented the regula falsi algorithm in LINESEARCH. This is a bounded root-finding algorithm that is guaranteed to converge if there there is a root in the interval. The LV sim had difficulty converging during the isovolumic phases, particularly as the pressure increased, but the simulation did not crash. Still, at many time steps during isovolumic phases, the residual was only ~1e-1. I think by allowing more line search iterations and more newton iterations, convergence will be better. Next steps are to not count line search iterations to Newton iterations, and make line search decrease the residual more than just 0.8 times the initial residual.